% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bmicat.R
\docType{data}
\name{data_bmicat}
\alias{data_bmicat}
\title{Sample data for body mass index (BMI) categorization.}
\format{
A tibble with 15 rows and 2 variables:
\describe{
\item{id}{\code{character}: Unique identifier for each individual}
\item{bmi}{\code{numeric}: Body Mass Index (BMI)}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("data_bmicat")
}
\description{
A dataset containing information about individuals, including their ID and
calculated Body Mass Index (BMI).
}
\examples{
dta_gtable(data_bmicat)

}
\keyword{datasets}
