% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cancer.R
\docType{data}
\name{data_cancer}
\alias{data_cancer}
\title{Sample data: Survival of cancer patients in days.}
\format{
A data frame with 4 rows and 18 variables:
\describe{
\item{cancer_type}{\code{character}: Type of cancer (e.g., "stomach",
"bronchus", "colon", "ovary").}
\item{V1}{\code{numeric}: Number of days.}
\item{V2}{\code{numeric}: Number of days.}
\item{V3}{\code{numeric}: Number of days.}
\item{V4}{\code{numeric}: Number of days.}
\item{V5}{\code{numeric}: Number of days.}
\item{V6}{\code{numeric}: Number of days.}
\item{V7}{\code{numeric}: Number of days.}
\item{V8}{\code{numeric}: Number of days.}
\item{V9}{\code{numeric}: Number of days.}
\item{V10}{\code{numeric}: Number of days.}
\item{V11}{\code{numeric}: Number of days.}
\item{V12}{\code{numeric}: Number of days.}
\item{V13}{\code{numeric}: Number of days.}
\item{V14}{\code{numeric}: Number of days.}
\item{V15}{\code{numeric}: Number of days.}
\item{V16}{\code{numeric}: Number of days.}
\item{V17}{\code{numeric}: Number of days.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("data_cancer")
}
\description{
A dataset containing information about the number of days patients survived
after being diagonized with level IV cancer. The cancer types are organized
by the rows. We will use this dataset to demonstrate transposing datasets.
}
\examples{
dta_gtable(data_cancer)

}
\keyword{datasets}
