% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gadgets.R
\docType{data}
\name{data_gadgets}
\alias{data_gadgets}
\title{Sample data: Gadgets owned by individuals}
\format{
A tibble with 15 rows and 1 variable:
\describe{
\item{gadgets_owned}{\code{character}: A comma-separated list of gadgets
owned by individuals.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("data_gadgets")
}
\description{
A dataset containing information about the gadgets owned by individuals.
Each entry lists one or more gadgets separated by commas.
}
\examples{
dta_gtable(data_gadgets)

}
\keyword{datasets}
