% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phone.R
\docType{data}
\name{data_phone}
\alias{data_phone}
\title{Sample data for phone satistifaction feedback.}
\format{
A tibble with 15 rows and 6 variables:
\describe{
\item{id}{\code{numeric}: Unique identifier for each review.}
\item{phone_type}{\code{character}: The phone model (OnePlus, Sumsung,
Google, Apple).}
\item{ease_of_use}{\code{character}: Rating for ease of use.}
\item{battery_life}{\code{character}: Rating for battery life.}
\item{camera_quality}{\code{character}: Rating for camera quality.}
\item{value_for_money}{\code{character}: Rating for value for money.}
\item{design_and_appearance}{\code{character}: Rating for design and appearance.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("data_phone")
}
\description{
A dataset containing reviews of various phone models, including user ratings
on ease of use, battery life, camera quality, value for money, and design
and appearance.
}
\examples{
dta_gtable(data_phone)

}
\keyword{datasets}
