% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_labels.R
\docType{data}
\name{dict_labels}
\alias{dict_labels}
\title{Dictionary with variable labels.}
\format{
A tibble with 3 rows and 2 variables:
\describe{
\item{names}{\code{character}: The variable name in the dataset.}
\item{labels}{\code{character}: A brief description of the variable.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("dict_labels")
}
\description{
A dictionary containing the names of variables and their descriptive labels.
}
\examples{
dta_gtable(dict_labels)

}
\keyword{datasets}
