% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_misspelled.R
\docType{data}
\name{dict_misspelled}
\alias{dict_misspelled}
\title{A dictionary with variable mappings for data standardization.}
\format{
A tibble with 29 rows and 3 variables:
\describe{
\item{variable}{\code{character}: The variable for which old and new
values are mapped.}
\item{old}{\code{character}: The old value to be replaced.}
\item{new}{\code{character}: The new value to replace the old value.}
}
}
\source{
Simulated data for demonstration purposes.
}
\usage{
data("dict_misspelled")
}
\description{
A dataset containing mappings of old values to new values for certain
specified variables in a dataset.
}
\examples{
dta_gtable(dict_misspelled)

}
\keyword{datasets}
