% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_rename.R
\docType{data}
\name{dict_rename}
\alias{dict_rename}
\title{A dictionary for variable name changes.}
\format{
A tibble with 9 rows and 2 variables:
\describe{
\item{old}{\code{character}: The old variable name.}
\item{new}{\code{character}: The new variable name that the old one has
been changed to.}
}
}
\source{
Simulated data for variable name mapping.
}
\usage{
data("dict_rename")
}
\description{
A dictionary containing mappings of old variable names to their
corresponding new names.
}
\examples{
dta_gtable(dict_rename)

}
\keyword{datasets}
