\name{Dm.tnumber}
\docType{data}
\alias{Dm.tnumber}
\title{The amount of thymines in the cDNA of each transcript of Drosophila Melanogaster.}
\description{
The amount of thymines in the cDNA of each transcript of all Drosophila Melanogaster Ensembl transcript IDs (Flybase transcript number), to assess the uridine-dependent labeling bias and eventually correct for it.
}
\usage{Dm.tnumber}
\format{
Vector gives the number of thymines in the cDNA (uridine residues in RNA) of each Ensembl transcript ID.
}
\source{
E. Birney, D. Andrews, M. Caccamo, Y. Chen, L. Clarke, G. Coates, T. Cox, F. Cunningham, V. Curwen, T. Cutts, T. Down, R. Durbin, X. M. Fernandez-Suarez, P. Flicek, S. Graef, M. Hammond, J. Herrero, K. Howe, V. Iyer, K. Jekosch, A. Kaehaeri, A. Kasprzyk, D. Keefe, F. Kokocinski, E. Kulesha, D. London, I. Longden, C. Melsopp, P. Meidl, B. Overduin, A. Parker, G. Proctor, A. Prlic, M. Rae, D. Rios, S. Redmond, M. Schuster, I. Sealy, S. Searle, J. Severin, G. Slater, D. Smedley, J. Smith, A. Stabenau, J. Stalker, S. Trevanion, A. Ureta-
Vidal, J. Vogel, S. White, C.Woodwark, and T. J. Hubbard. Ensembl 2006. Nucleic acids research, 34(Database issue), January 2006.
}
\keyword{datasets}
