\name{Doelken2008}
\docType{data}
\alias{Doelken2008}
\title{Mus Musculus and Homo Sapiens DTA experiment from Doelken et al.}
\description{
R object contains all relevant *.RData files needed for the \code{DTA.estimate} function. For example, see vignette.
}
\usage{Doelken2008}
\format{
R object contains the following *.RData files:
\code{Hs.phenomat}
\code{Hs.datamat}
\code{Hs.reliable}
\code{Hs.enst2ensg}
\code{Hs.tnumber}
\code{Mm.phenomat}
\code{Mm.datamat}
\code{Mm.reliable}
\code{Mm.enst2ensg}
\code{Mm.tnumber}
}
\source{
Doelken, L., Ruzsics, Z., Raedle, B., Friedel, C. C., Zimmer, R., Mages, J., Hoffmann, R., Dickinson, P., Forster, T., Ghazal, P., & Koszinowski, U. H. (2008). High-resolution gene expression profiling for simultaneous kinetic parameter analysis of RNA synthesis and decay. RNA 14(9), 1959-1972.
E. Birney, D. Andrews, M. Caccamo, Y. Chen, L. Clarke, G. Coates, T. Cox, F. Cunningham, V. Curwen, T. Cutts, T. Down, R. Durbin, X. M. Fernandez-Suarez, P. Flicek, S. Graef, M. Hammond, J. Herrero, K. Howe, V. Iyer, K. Jekosch, A. Kaehaeri, A. Kasprzyk, D. Keefe, F. Kokocinski, E. Kulesha, D. London, I. Longden, C. Melsopp, P. Meidl, B. Overduin, A. Parker, G. Proctor, A. Prlic, M. Rae, D. Rios, S. Redmond, M. Schuster, I. Sealy, S. Searle, J. Severin, G. Slater, D. Smedley, J. Smith, A. Stabenau, J. Stalker, S. Trevanion, A. Ureta-
Vidal, J. Vogel, S. White, C.Woodwark, and T. J. Hubbard. Ensembl 2006. Nucleic acids research, 34(Database issue), January 2006.
}
\keyword{datasets}
