\name{DTA-package}
\alias{DTA-package}
\alias{DTA}
\docType{package}
\title{Dynamic Transcriptome Analysis}
\description{
The \code{DTA} package implements all methods of the quantitative kinetic modeling approach belonging to DTA (Dynamic Transcriptome Analysis) to estimate mRNA synthesis and decay rates from individual time point measurements.
}
\author{Bjoern Schwalb <schwalb@lmb.uni-muenchen.de>}
\references{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Dolken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
}
\keyword{package}
\examples{
## see vignette
}
