\name{DTA.map.it}
\alias{DTA.map.it}
\title{Mapping function to switch between different identifiers.}
\description{\code{DTA.map.it} can map different kinds of identifiers in a matrix or a vector given by mapping vector.}
\usage{
DTA.map.it(mat,map = NULL,check = TRUE) 
}
\arguments{
\item{mat}{Matrix or vector with numerical entries.}
\item{map}{Vector of identifiers to map to, named by identifiers to map from.}
\item{check}{Should check protocol be printed.}
}
\author{Bjoern Schwalb \email{schwalb@lmb.uni-muenchen.de}}
\examples{
### see vignette examples or reference:
### B. Schwalb, B. Zacher, S. Duemcke, D. Martin, P. Cramer, A. Tresch.
### Measurement of genome-wide RNA synthesis and decay rates with Dynamic Transcriptome Analysis (DTA/cDTA). Bioinformatics, in revision.
}
\references{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Genome Research, in revision.
B. Schwalb, B. Zacher, S. Duemcke, D. Martin, P. Cramer, A. Tresch. Measurement of genome-wide RNA synthesis and decay rates with Dynamic Transcriptome Analysis (DTA/cDTA). Bioinformatics, in revision. 
}
\keyword{utilities}

