% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_bmi.R
\name{dta_bmi}
\alias{dta_bmi}
\title{Calculate body mass index (BMI)}
\usage{
dta_bmi(
  dat,
  .weight,
  .height,
  weight_units = c("kg", "lbs"),
  height_units = c("m", "in"),
  name = "bmi",
  digits = NULL
)
}
\arguments{
\item{dat}{A data frame or tibble containing the weight and height columns.}

\item{.weight}{Name of the column representing weight in the data.}

\item{.height}{Name of the column representing height in the data.}

\item{weight_units}{Character string specifying the units of weight.
Options are \code{"kg"} for kilograms or \code{"lbs"}
for pounds. Default is \code{"kg"}.}

\item{height_units}{Character string specifying the units of height.
Options are \code{"m"} for meters or \code{"in"} for
inches. Default is \code{"m"}.}

\item{name}{Character string specifying the name of the new BMI column.
Default is \code{bmi}.}

\item{digits}{Integer specifying the number of decimal places for rounding
the BMI value. Default is \code{NULL}.}
}
\value{
A tibble with the calculated body mass index (BMI).
}
\description{
\code{dta_bmi()} computes the Body Mass Index (BMI) given weight and height
columns in a data frame or tibble. It allows customization for weight and
height units.
}
\examples{
data("data_bmi")
dta_gtable(data_bmi)

# Calculate BMI from the columns `weight` and `height`

df <- dta_bmi(
  dat = data_bmi,
  .weight = weight,
  .height = height
)
dta_gtable(df)

# Calculate BMI from the columns `weight` and `height`
# to 2 decimal points and assign the values to the new
# variable named `body_mass_index`

df2 <- dta_bmi(
  dat = data_bmi,
  .weight = weight,
  .height = height,
  name = body_mass_index,
  digits = 2
)
dta_gtable(df2)

}
\references{
\url{https://www.cdc.gov/bmi/adult-calculator/bmi-categories.html}
}
