% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_bmi.R
\name{dta_bmicat}
\alias{dta_bmicat}
\title{Categorize BMI into weight categories}
\usage{
dta_bmicat(
  dat,
  .bmi,
  name = "bmi_cat",
  is_extended = FALSE,
  out_of_range = "",
  as_factor = TRUE,
  as_numeric = FALSE
)
}
\arguments{
\item{dat}{A data frame or tibble containing the BMI data.}

\item{.bmi}{The name of the column in \code{dat} containing BMI values.}

\item{name}{A character string specifying the name of the new column to be
created with BMI categories. Default is \code{bmi_cat}.}

\item{is_extended}{A logical value indicating whether to use the extended
BMI categories. If \code{TRUE}, includes Class 1, 2,
and 3 obesity. Default is \code{FALSE}. See Details
section for more information.}

\item{out_of_range}{A character string specifying the label for out-of-range
BMI values. Default is an empty string.}

\item{as_factor}{A logical value indicating whether to return the BMI
categories as a factor. Default is \code{TRUE}.}

\item{as_numeric}{A logical value indicating whether to convert the factor
to numeric. Only applicable if \code{as_factor} is
\code{TRUE}. Default is \code{FALSE}.}
}
\value{
A data frame or tibble with an additional column for BMI categories.
}
\description{
\code{dta_bmicat()} categorizes Body Mass Index (BMI) values into specified
weight categories.
}
\details{
The standard BMI categories are defined as follows:

\describe{
\item{Underweight}{BMI < 18.5}
\item{Healthy weight}{18.5 ≤ BMI < 25.0}
\item{Overweight}{25.0 ≤ BMI < 30.0}
\item{Obesity}{BMI ≥ 30.0}
}

If \code{is_extended} is set to \code{TRUE}, the extended BMI categories are:

\describe{
\item{Underweight}{BMI < 18.5}
\item{Healthy weight}{18.5 ≤ BMI < 25.0}
\item{Overweight}{25.0 ≤ BMI < 30.0}
\item{Class I Obesity}{30.0 ≤ BMI < 35.0}
\item{Class II Obesity}{35.0 ≤ BMI < 40.0}
\item{Class III Obesity (Severe)}{BMI ≥ 40.0}
}
}
\examples{
data("data_bmicat")
dta_gtable(data_bmicat)

# Categorize `bmi` into the standard BMI categories

df <- dta_bmicat(
  dat = data_bmicat,
  .bmi = bmi,
  name = bmi_cat,
  is_extended = FALSE,
  as_factor = TRUE
)
dta_gtable(df)

# Categorize `bmi` into the extended BMI categories

df2 <- dta_bmicat(
  dat = data_bmicat,
  .bmi = bmi,
  name = bmi_cat,
  is_extended = TRUE,
  as_factor = TRUE
)
dta_gtable(df2)

# Categorize `bmi` into the standard BMI categories and
# convert to numeric

df3 <- dta_bmicat(
  dat = data_bmicat,
  .bmi = bmi,
  name = bmi_cat,
  is_extended = TRUE,
  as_factor = TRUE,
  as_numeric = TRUE
)
dta_gtable(df3)

}
\seealso{
\code{\link{dta_bmi}}, \code{\link[dplyr]{case_match}}, \code{\link[dplyr]{case_when}}
}
