% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_columns.R
\name{dta_columns}
\alias{dta_columns}
\title{Retrieve column names from a data frame}
\usage{
dta_columns(dat, .columns, is_one_column = FALSE)
}
\arguments{
\item{dat}{A data frame or tibble from which to select columns.}

\item{.columns}{Column name(s) to be selected from \code{dat}. Supports
tidy selection syntax.}

\item{is_one_column}{A logical value indicating whether only one column
should be selected. If \code{TRUE} and more than one
column is selected, an error is thrown. Default is
\code{FALSE}.}
}
\value{
A character vector containing the names of the selected columns.
}
\description{
\code{dta_columns()} retrieves the names of specified columns from
a data frame or tibble.
}
\examples{
data(mtcars)
dta_gtable(head(mtcars))

dta_columns(mtcars, .columns = starts_with("c"))

dta_columns(mtcars, .columns = cyl:wt)

dta_columns(mtcars, .columns = c(mpg, hp, vs, gear))

try(dta_columns(mtcars, .columns = c(mpg, hp, vs, gear), is_one_column = TRUE))

}
