% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_crosstab.R
\name{dta_crosstab}
\alias{dta_crosstab}
\title{Generate cross tabulations with optional percentages and totals}
\usage{
dta_crosstab(
  dat,
  .row,
  .column,
  cells = c("counts", "row", "col"),
  add_totals = c("both", "row", "col"),
  name = "Variable",
  add_percent_symbol = TRUE,
  digits = 2
)
}
\arguments{
\item{dat}{A data frame (not a tibble) containing the variables to tabulate.}

\item{.row}{The variable to be used as rows in the crosstab.}

\item{.column}{The variable to be used as columns in the crosstab.}

\item{cells}{A character string indicating the type of values to display in
the crosstab:
\describe{
\item{\code{"counts"}}{Display counts (default).}
\item{\code{"row"}}{Display row percentages.}
\item{\code{"col"}}{Display column percentages.}
}}

\item{add_totals}{A character string specifying where to add totals:
\describe{
\item{\code{"both"}}{Add totals to both rows and columns (default).}
\item{\code{"row"}}{Add totals to rows only.}
\item{\code{"col"}}{Add totals to columns only.}
}}

\item{name}{A character string to rename the first column in the output.
Default is \code{"Variable"}.}

\item{add_percent_symbol}{Logical indicating whether or not to add a \% sign
to percentages. Default is \code{TRUE}.}

\item{digits}{An integer specifying the number of decimal places to use for
percentages. Default is 1.}
}
\value{
A data frame representing the cross-tabulation, styled using
\code{janitor} functions.
The format of the output depends on the \code{cells} argument:
\describe{
\item{Counts}{Displays raw counts.}
\item{Percentages}{Displays percentages with counts in parentheses if \code{cells} is \code{"row"} or \code{"col"}.}
}
}
\description{
\code{dta_crosstab()} creates a cross-tabulation (contingency table) for
a given dataset, with options to include counts, row percentages, or column
percentages. Totals can also be added to rows, columns, or both. The output
is styled using the \code{janitor} package.
}
\examples{
data("data_sample")
df <- data_sample

# Crosstabulation of frequencies (counts)

result <- dta_crosstab(
  dat = df, .row = region, .column = age_group
)
dta_gtable(result)

# Calculate column percentages

result2 <- dta_crosstab(
  dat = df, 
  .row = region,
  .column = age_group,
  cells = "col",
  add_totals = "col"
)
dta_gtable(result2)

# Calculate row percentages

result3 <- dta_crosstab(
  dat = df,
  .row = region,
  .column = age_group,
  cells = "row",
  add_totals = "row"
)
dta_gtable(result3)


# Remove the percentages symbol

result4 <- dta_crosstab(
  dat = df,
  .row = region,
  .column = age_group,
  cells = "row",
  add_totals = "row",
  add_percent_symbol = FALSE
)
dta_gtable(result4)

}
