% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_drop_na_with_warning.R
\name{dta_drop_with_warning}
\alias{dta_drop_with_warning}
\title{Drop rows with missing values and issue a warning}
\usage{
dta_drop_with_warning(dat, .columns = names(dat))
}
\arguments{
\item{dat}{A data frame or tibble from which rows with missing values
should be dropped.}

\item{.columns}{Columns to check for missing values. Supports tidy
selection syntax. Defaults to \code{NULL}, meaning
all columns are checked.}
}
\value{
A data frame or tibble with rows containing missing values in the
specified columns removed.
}
\description{
\code{dta_drop_with_warning()} removes rows with missing values (NA) from
specified columns in a data frame or tibble. If any rows are dropped,
a warning is issued indicating the number of rows removed.
}
\examples{
df <- data.frame(x = c(1, 2, NA, 4), y = c(NA, 5, 6, 7))
dta_gtable(df)

# Drop rows with NA values in any column and issue a warning

dta_drop_with_warning(df)

# Drop rows with NA values in a specific column and issue a warning

dta_drop_with_warning(df, .columns = x)

}
\seealso{
\code{\link[tidyr]{drop_na}}
}
