% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_freq.R
\name{dta_freq}
\alias{dta_freq}
\title{Get frequency distribution of a specific column}
\usage{
dta_freq(
  dat,
  .column,
  name = NULL,
  is_sorted = FALSE,
  is_decreasing = TRUE,
  add_percent_symbol = TRUE,
  digits = 2
)
}
\arguments{
\item{dat}{A dataframe.}

\item{.column}{The column in the dataframe for which the frequency
distribution will be calculated.}

\item{name}{Optional. A name for the column in the output. Defaults to
\code{NULL}.}

\item{is_sorted}{Logical. If \code{TRUE}, sorts the results. Defaults to
\code{FALSE}.}

\item{is_decreasing}{Logical. If \code{TRUE}, sorts in decreasing order.
Defaults to \code{TRUE}.}

\item{add_percent_symbol}{Logical indicating whether or not to add a \% sign
to percentages. Default is \code{TRUE}.}

\item{digits}{Integer. Number of digits to round the percentage values
to. Defaults to \code{1}.}
}
\value{
A dataframe with frequency counts, percentages, and total for the
specified column.
}
\description{
\code{dta_freq()} calculates the frequency distribution of a specified
column in a dataframe, with options for sorting, showing missing values,
and customizing the output.
}
\examples{
# Basic frequency distribution of 'region' column

data("data_sample")
tab <- dta_freq(dat = data_sample, .column = region)
dta_gtable(tab)

# Frequency distribution with sorting in increasing
# order of frequency

tab2 <- dta_freq(
  dat = data_sample,
  .column = region,
  is_sorted = TRUE,
  is_decreasing = FALSE
)
dta_gtable(tab2)

# Frequency distribution with sorting in decreasing
# order of frequency

tab3 <- dta_freq(
  dat = data_sample,
  .column = region,
  is_sorted = TRUE,
  is_decreasing = TRUE
)
dta_gtable(tab3)

# Remove the percentage symbol

tab3 <- dta_freq(
  dat = data_sample,
  .column = region,
  is_sorted = TRUE,
  is_decreasing = TRUE,
  add_percent_symbol = FALSE
)
dta_gtable(tab3)

}
