% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dta_get_datasets}
\alias{dta_get_datasets}
\title{Export the sample datasets to file}
\usage{
dta_get_datasets(copy_to_file_path = NULL)
}
\arguments{
\item{copy_to_file_path}{A string specifying the file path where the
dataset should be copied. If \code{NULL} (default),
the dataset will be copied to \code{dta-data.xlsx}
in the current working directory.}
}
\value{
The function does not return a value. It prints a success message
if the operation is successful or throws an error if it fails.
}
\description{
\code{dta_get_datasets()} copies a dataset from the \code{code} installation
to a specified file path. If no path is provided, it saves the file as
\code{dta-data.xlsx} in the current working directory.
}
\details{
The function uses \code{file.copy()} to copy the file specified by
\code{dta_path()} to the destination path. The destination path can be
provided as an argument, or it defaults to \code{dta-data.xlsx} in the
current working directory.
}
\examples{
# Copy to the default location
# dta_get_datasets()

}
\seealso{
\code{\link{dta_path}} \code{\link[base]{file.copy}}, \code{\link[base]{getwd}}
}
