% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_gt.R
\name{dta_gt}
\alias{dta_gt}
\title{Create a styled gt table}
\usage{
dta_gt(dat)
}
\arguments{
\item{dat}{A data frame that you want to render and style as a \code{gt}
table.}
}
\value{
A styled \code{gt} table object.
}
\description{
\code{dta_gt()} creates a \code{gt} table from the input data frame and
applies a consistent styling with vertical borders, bold column labels,
reduced padding, and customized background colors.
}
\examples{
data("women")
dta_gt(women)

data("mtcars")
dta_gt(head(mtcars, n = 10))

}
