% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dta_path}
\alias{dta_path}
\title{Get the path to the example dataset}
\usage{
dta_path()
}
\value{
A string containing the full file path to the dataset within the
package. If the file does not exist, it returns an empty string.
}
\description{
\code{dta_path()} provides the file path to the example dataset
\code{dta-data.xlsx} included in the package. The dataset is located in the
\code{extdata} directory of the package.
}
\details{
The function uses \code{\link[base]{system.file}} to construct the path to
the \code{extdata/dta-data.xlsx} file included in the package. This is
useful for accessing built-in example files.
}
\examples{
# Get the file path to the dataset
# dta_path()

}
\seealso{
\code{\link[base]{system.file}}
}
