% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_recode.R
\name{dta_recode}
\alias{dta_recode}
\title{Recode variables in a data frame based on a dictionary}
\usage{
dta_recode(
  dat,
  dict,
  sheet = 1,
  min_categories = 2,
  max_categories = 25,
  as_numeric = FALSE,
  is_force_sequential = FALSE
)
}
\arguments{
\item{dat}{A data frame or tibble containing the variables to be recoded.}

\item{dict}{A data frame or tibble serving as the dictionary, specifying
variable names, values, and labels.}

\item{sheet}{The name or index of the worksheet that contains the data
for the dictionary.}

\item{min_categories}{Minimum number of categories for a variable to be
recoded. Defaults to \code{1}.}

\item{max_categories}{Maximum number of categories for a variable to be
recoded. Defaults to \code{25}.}

\item{as_numeric}{Logical. If \code{TRUE}, the recoded variables are
returned as numeric. Defaults to \code{FALSE}.}

\item{is_force_sequential}{Logical indicating whether or not to force
sequential values, that is, they should start
at 1 and increase by 1.}
}
\value{
A tibble with recoded variables. If warnings are generated,
they are saved to a CSV file and displayed.
}
\description{
\code{dta_recode()} recodes variables in a \code{dat} using a dictionary
\code{dict}. The dictionary maps the original values of each variable to
their new values and labels.
}
\examples{
library(dplyr)
data("data_sample")
glimpse(data_sample) # look at the data type column

data("dict_recode")
dta_gtable(dict_recode)

# The default nature of `dta_recode()` is to drop the
# labels if the values are not sequential or do not
# start at 1. To maintain these labels, set 
# `is_force_sequential` to `TRUE`. Note that this will
# reset the given values to sequential.

result2 <- dta_recode(
  dat = data_sample,
  dict = dict_recode,
  is_force_sequential = TRUE
)
glimpse(result2)

# Return numeric codes

result3 <- dta_recode(
  dat = data_sample,
  dict = dict_recode,
  as_numeric = TRUE
)
glimpse(result3) # look at the data type column and values

}
\seealso{
\code{\link{dta_recode_common_labels}}, \code{\link{dta_recode_auto}}, \code{\link[dplyr]{case_match}}, \code{\link[dplyr]{case_when}}
}
