% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_recode_auto.R
\name{dta_recode_auto}
\alias{dta_recode_auto}
\title{Automatically recode categorical variables in a data frame}
\usage{
dta_recode_auto(
  dat,
  .columns = names(dat),
  min_categories = 2,
  max_categories = 25,
  is_ordered = FALSE,
  as_numeric = FALSE
)
}
\arguments{
\item{dat}{A data frame containing the variables to be recoded.}

\item{.columns}{A character vector specifying which columns to recode.
Default is \code{NULL}, meaning all applicable columns will
be recoded.}

\item{min_categories}{The minimum number of unique values to consider for
recoding to happen. Default is \code{2}.}

\item{max_categories}{The maximum number of unique values to consider afor
recoding to happen. Default is \code{25}.}

\item{is_ordered}{Logical. If \code{TRUE}, creates an ordered factor.
Default is \code{FALSE}.}

\item{as_numeric}{Logical. If \code{TRUE}, converts the result of recoding
to numeric, otherwise, it keeps them as factors. Default
is \code{FALSE}.}
}
\value{
A data frame with recoded categorical variables.
}
\description{
\code{dta_recode_auto()} automatically recodes categorical variables in
a data frame or tibble. That is, it assigns numerical codes starting
from 1 and increasing by 1 (i.e. 1, 2, 3, 4, ...) to the labels
(categories) according to their alphabetic order.
}
\details{
The function automatically recodes categorical variables in the data frame
that have a number of unique values between \code{min_categories} and
\code{max_categories}. It assigns a numeric code to each unique category
based on their alphabetical order. For instance, if a column represents
the variable \code{Gender} with categories \code{Female} and \code{Male},
the function assigns \code{1} to \code{Female} and \code{2} to \code{Male}.
This recoding is done using the \code{factor} function, where the levels of
the factor are ordered alphabetically, and then converted to numeric values.
This ensures that the recoding is consistent and follows a lexicographical
order for the categorical variables.
}
\examples{
library(dplyr)
data("data_sample")
glimpse(data_sample) # look at the data type column

# Auto-recode all categorical variables

result <- dta_recode_auto(dat = data_sample)
glimpse(result) # look at the data type column

# Convert to numeric

result <- dta_recode_auto(
  dat = data_sample, as_numeric = TRUE
)
glimpse(result) # look at the data type and values columns

# Specify the columns using the `.columns` argument

result2 <- dta_recode_auto(
  dat = data_sample, .columns = r:excel
)
glimpse(result2) # only the variables `r` through `excel` will be recoded

}
\seealso{
\code{\link{dta_recode}}, \code{\link{dta_recode_common_labels}}, \code{\link[dplyr]{case_match}}, \code{\link[dplyr]{case_when}}
}
