% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_replace.R
\name{dta_replace}
\alias{dta_replace}
\title{Correct misspelled data using a dictionary}
\usage{
dta_replace(dat, dict, .name, .wrong, .correct)
}
\arguments{
\item{dat}{A data frame or tibble containing the data to be corrected.}

\item{dict}{A data frame or tibble serving as the dictionary, with columns
specifying the correct and incorrect spellings.}

\item{.name}{The column in both \code{dat} and \code{dict} to match entries
by (e.g., a unique identifier).}

\item{.wrong}{The column in \code{dict} containing the misspelled values to
be corrected.}

\item{.correct}{The column in \code{dict} containing the correct values for
the misspelled entries.}
}
\value{
A data frame or tibble with corrected entries.
}
\description{
\code{dta_replace()} corrects misspelled entries in a data frame or tibble
based on a provided dictionary. The dictionary specifies the correct values
for misspelled entries in a specified column.
}
\details{
The function first validates that \code{dat} and \code{dict} are data frames
or tibbles. It then fills missing values in the \code{dict} for the columns
specified in \code{.name} and \code{.correct}, using a downward fill
strategy. Finally, it replaces misspelled values in \code{dat} using a
dictionary lookup facilitated by \code{matchmaker::match_df()}.
}
\examples{
# Example data with misspelled characters / strings

data("data_misspelled")
dta_gtable(head(data_misspelled))

data("dict_misspelled")
dta_gtable(dict_misspelled)

# Correct the misspelled entries in `dat` using the
# `dict` dictionary

result <- dta_replace(
  dat = data_misspelled, 
  dict = dict_misspelled, 
  .name = variable, 
  .wrong = old, 
  .correct = new
)
dta_gtable(head(result))

}
