% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dta_to_numeric.R
\name{dta_to_numeric}
\alias{dta_to_numeric}
\title{Convert numeric strings in a data frame or tibble to numeric numbers}
\usage{
dta_to_numeric(dat)
}
\arguments{
\item{dat}{A data frame where numeric strings in character columns should
be converted to numeric.}
}
\value{
A data frame with numeric strings converted to numeric columns.
}
\description{
\code{dta_to_numeric()} takes a data frame and converts columns containing
numeric values represented as strings into numeric columns. Other columns
remain unchanged.
}
\details{
The function scans each column in the input data frame. For character
columns, it checks if all the values can be safely coerced to numeric.
If so, the column is converted to numeric using \code{as.numeric()}.
Non-character columns or character columns containing non-numeric values
remain unchanged.

Columns are processed using \code{lapply()}, which ensures efficient
handling of each column.
}
\examples{
# A data frame with numeric character (a), characters (b) and numeric numbers (c)

df <- data.frame(
  a = c("1", "2", "3"),
  b = c("A", "B", "C"),
  c = c(4, 5, 6)
)
dta_gtable(df)
str(df)

# Convert numeric strings to numeric columns

df <- dta_to_numeric(df)
dta_gtable(df)
str(df)

}
