\name{Hs.datamat}
\docType{data}
\alias{Hs.datamat}
\title{Gene expression profiles of the Homo Sapiens DTA experiment from Doelken et al.}
\description{
This matrix contains the RNA intensity values for each gene across each RNA fraction and their replicate measurements of the Homo Sapiens DTA experiment from Doelken et al.
}
\usage{Hs.datamat}
\format{
The column names of the matrix give the cel-file name and the row names the Ensembl gene IDs.
}
\source{
Doelken, L., Ruzsics, Z., Raedle, B., Friedel, C. C., Zimmer, R., Mages, J., Hoffmann, R., Dickinson, P., Forster, T., Ghazal, P., & Koszinowski, U. H. (2008). High-resolution gene expression profiling for simultaneous kinetic parameter analysis of RNA synthesis and decay. RNA 14(9), 1959-1972.
}
\keyword{datasets}
