\name{Miller2011dynamic}
\docType{data}
\alias{Miller2011dynamic}
\title{Saccharomyces Cerevisiae salt stress DTA experiment from Miller et al.}
\description{
R object contains all relevant *.RData files needed for the \code{DTA.estimate} function. For example, see vignette.
}
\usage{Miller2011dynamic}
\format{
R object contains the following *.RData files:
\code{Sc.phenomat.dynamic}
\code{Sc.datamat.dynamic}
\code{Sc.reliable.dynamic}
\code{Sc.tnumber}
}
\source{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Doelken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
E. Birney, D. Andrews, M. Caccamo, Y. Chen, L. Clarke, G. Coates, T. Cox, F. Cunningham, V. Curwen, T. Cutts, T. Down, R. Durbin, X. M. Fernandez-Suarez, P. Flicek, S. Graef, M. Hammond, J. Herrero, K. Howe, V. Iyer, K. Jekosch, A. Kaehaeri, A. Kasprzyk, D. Keefe, F. Kokocinski, E. Kulesha, D. London, I. Longden, C. Melsopp, P. Meidl, B. Overduin, A. Parker, G. Proctor, A. Prlic, M. Rae, D. Rios, S. Redmond, M. Schuster, I. Sealy, S. Searle, J. Severin, G. Slater, D. Smedley, J. Smith, A. Stabenau, J. Stalker, S. Trevanion, A. Ureta-
Vidal, J. Vogel, S. White, C.Woodwark, and T. J. Hubbard. Ensembl 2006. Nucleic acids research, 34(Database issue), January 2006.
}
\keyword{datasets}
