\name{Mm.enst2ensg}
\docType{data}
\alias{Mm.enst2ensg}
\title{Mapping of Mus Musculus gene and transcript identifiers.}
\description{
Mapping from Ensembl transcript IDs to Ensembl gene IDs of Mus Musculus.
}
\usage{Mm.enst2ensg}
\format{
Vector gives the Ensembl gene IDs, names the Ensembl transcript IDs.
}
\source{
E. Birney, D. Andrews, M. Caccamo, Y. Chen, L. Clarke, G. Coates, T. Cox, F. Cunningham, V. Curwen, T. Cutts, T. Down, R. Durbin, X. M. Fernandez-Suarez, P. Flicek, S. Graef, M. Hammond, J. Herrero, K. Howe, V. Iyer, K. Jekosch, A. Kaehaeri, A. Kasprzyk, D. Keefe, F. Kokocinski, E. Kulesha, D. London, I. Longden, C. Melsopp, P. Meidl, B. Overduin, A. Parker, G. Proctor, A. Prlic, M. Rae, D. Rios, S. Redmond, M. Schuster, I. Sealy, S. Searle, J. Severin, G. Slater, D. Smedley, J. Smith, A. Stabenau, J. Stalker, S. Trevanion, A. Ureta-
Vidal, J. Vogel, S. White, C.Woodwark, and T. J. Hubbard. Ensembl 2006. Nucleic acids research, 34(Database issue), January 2006.
}
\keyword{datasets}
