\name{Mm.phenomat}
\docType{data}
\alias{Mm.phenomat}
\title{Design of the Mus Musculus DTA experiment from Doelken et al.}
\description{
The phenotype matrix \code{Mm.phenomat} contains information about the experimental design. It is comprised of the filename, the type of RNA fraction measured (T, U or L), the labeling time and the replicate number.
}
\usage{Mm.phenomat}
\format{
The phenomat is a matrix comprised of the file name, the type of RNA fraction mesasured (T, U or L, \code{fraction} column), the labeling time (\code{time},\code{timeframe} column) and the replicate number (\code{nr} column). Rows in this matrix represent the individual experiments.
}
\source{
Doelken, L., Ruzsics, Z., Raedle, B., Friedel, C. C., Zimmer, R., Mages, J., Hoffmann, R., Dickinson, P., Forster, T., Ghazal, P., & Koszinowski, U. H. (2008). High-resolution gene expression profiling for simultaneous kinetic parameter analysis of RNA synthesis and decay. RNA 14(9), 1959-1972.
}
\keyword{datasets}
