\name{Mm.reliable}
\docType{data}
\alias{Mm.reliable}
\title{Gene identifiers valid for parameter estimation from the Mus Musculus Doelken et al. DTA experiment.}
\description{
Ensembl gene IDs, that passed certain criteria among the Mus Musculus Doelken et al. DTA experiment to be considered valid for parameter estimation. For details, see vignette.
}
\usage{Mm.reliable}
\format{
Vector of Ensembl gene IDs that can be passed to \code{DTA.estimate} for parameter estimation.
}
\source{
Doelken, L., Ruzsics, Z., Raedle, B., Friedel, C. C., Zimmer, R., Mages, J., Hoffmann, R., Dickinson, P., Forster, T., Ghazal, P., & Koszinowski, U. H. (2008). High-resolution gene expression profiling for simultaneous kinetic parameter analysis of RNA synthesis and decay. RNA 14(9), 1959-1972.
}
\keyword{datasets}
