\name{orfs_ribig}
\docType{data}
\alias{orfs_ribig}
\title{Ribosome biogenesis genes}
\description{
ORF identifiers found to be associated with ribosome biogenesis, rRNA processing etc.
}
\usage{orfs_ribig}
\format{
Vector of ORF identifiers.
}
\source{
P. Jorgensen, I. RupeAi, J. R. Sharom, L. Schneper, J. R. Broach, and M. Tyers. A dynamic
transcriptional network communicates growth potential to ribosome synthesis and critical cell
size. Genes & Development, 18(20):2491-2505, October 2004.
}
\keyword{datasets}
