\name{orfs_rpg}
\docType{data}
\alias{orfs_rpg}
\title{Ribosomal protein genes}
\description{
ORF identifiers encoding for ribosomal protein genes.
}
\usage{orfs_rpg}
\format{
Vector of ORF identifiers.
}
\source{
A. Nakao, M. Yoshihama, and N. Kenmochi. RPG: the Ribosomal Protein Gene database.
Nucleic acids research, 32(Database issue), January 2004.
}
\keyword{datasets}
