\name{orfs_stress}
\docType{data}
\alias{orfs_stress}
\title{ISA stress module}
\description{
ORF identifiers found to be associated with stress response by the iterative signature algorithm.
}
\usage{orfs_stress}
\format{
Vector of ORF identifiers.
}
\source{
J. Ihmels, G. Friedlander, S. Bergmann, O. Sarig, Y. Ziv, and N. Barkai. Revealing modular
organization in the yeast transcriptional network. Nature genetics, 31(4):370-377, August 2002.
}
\keyword{datasets}
