\name{orfs_tf}
\docType{data}
\alias{orfs_tf}
\title{Transcription factors}
\description{
ORF identifiers encoding for transcription factors.
}
\usage{orfs_tf}
\format{
Vector of ORF identifiers.
}
\source{
K. D. MacIsaac, T. Wang, D. B. Gordon, D. K. Gifford, G. D. Stormo, and E. Fraenkel. An
improved map of conserved regulatory sites for saccharomyces cerevisiae. BMC Bioinformatics,
7:113, 2006.
}
\keyword{datasets}
