\name{phenomat}
\docType{data}
\alias{phenomat}
\title{DTA phenomat}
\description{
Design of a wild-type yeast DTA experiment.
}
\usage{phenomat}
\format{
The phenomat is comprised of the file name, the type of RNA fraction mesasured (T, U or L), the labeling time and the replicate number. Rows in this matrix represent the individual experiments.
}
\source{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Dolken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
}
\keyword{datasets}
