\name{Raw.datamat}
\docType{data}
\alias{Raw.datamat}
\title{Gene expression profiles of the Saccharomyces Cerevisiae rpb1-N488D (Slow Polymerase) and wild-type cDTA experiment from Sun et al.}
\description{
This matrix contains the RNA intensity values for each gene across each RNA fraction and their replicate measurements of the Saccharomyces Cerevisiae rpb1-N488D (Slow Polymerase) and wild-type cDTA experiment from Sun et al.
}
\usage{Raw.datamat}
\format{
The column names of the matrix give the cel-file name and the row names the affymetrix IDs.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Under review.
}
\keyword{datasets}
