\name{reliable}
\docType{data}
\alias{reliable}
\title{DTA reliable ORFs}
\description{
4490 genes, that passed certain criteria to be considered valid for parameter estimation (see vignette for details).
}
\usage{reliable}
\format{
Vector of ORF identifiers.
}
\source{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Dolken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
}
\keyword{datasets}
