\name{RPphenomat}
\docType{data}
\alias{RPphenomat}
\title{cDTA RPphenomat}
\description{
Design of a rpb1-N488D (Slow Polymerase) cDTA experiment.
}
\usage{RPphenomat}
\format{
The phenomat is comprised of the file name, the type of RNA fraction mesasured (T, U or L), the labeling time and the replicate number. Rows in this matrix represent the individual experiments.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. submitted.
}
\keyword{datasets}
