\name{Sc.affy2ensg}
\docType{data}
\alias{Sc.affy2ensg}
\title{Mapping of SaccharomycesCerevisiae Affymetrix Yeast 2.0 and gene identifiers.}
\description{
Mapping from Affymetrix Yeast 2.0 IDs to Ensembl gene IDs of SaccharomycesCerevisiae.
}
\usage{Sc.affy2ensg}
\format{
Vector gives the Ensembl gene IDs, names the Affymetrix Yeast 2.0 IDs.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Genome Research, in revision.
}
\keyword{datasets}
