\name{Sc.ensg.reliable}
\docType{data}
\alias{Sc.ensg.reliable}
\title{Gene identifiers valid for parameter estimation from the Saccharomyces Cerevisiae Sun et al. cDTA experiment.}
\description{
Ensembl gene IDs, that passed certain criteria among the Saccharomyces Cerevisiae Sun et al. cDTA experiment to be considered valid for parameter estimation. For details, see Sun et al (Materials and Methods).
}
\usage{Sc.ensg.reliable}
\format{
Vector of Ensembl gene IDs that can be passed to \code{DTA.estimate} for parameter estimation.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Under review.
}
\keyword{datasets}
