\name{Sc.reliable.dynamic}
\docType{data}
\alias{Sc.reliable.dynamic}
\title{Gene identifiers valid for parameter estimation from the Saccharomyces Cerevisiae Miller et al. salt stress DTA experiment.}
\description{
Ensembl gene IDs, that passed certain criteria among the Saccharomyces Cerevisiae Miller et al. salt stress DTA experiment to be considered valid for parameter estimation. For details, see supplemental material Miller et al.
}
\usage{Sc.reliable.dynamic}
\format{
Vector of Ensembl gene IDs that can be passed to \code{DTA.estimate} for parameter estimation.
}
\source{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Doelken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
}
\keyword{datasets}
