\name{Sc.ribig.ensg}
\docType{data}
\alias{Sc.ribig.ensg}
\title{Ribosome biogenesis genes.}
\description{
ORF identifiers (Ensembl Gene ID) found to be associated with ribosome biogenesis, rRNA processing etc.
}
\usage{Sc.ribig.ensg}
\format{
Vector of ORF identifiers (Ensembl Gene ID).
}
\source{
P. Jorgensen, I. RupeAi, J. R. Sharom, L. Schneper, J. R. Broach, and M. Tyers. A dynamic
transcriptional network communicates growth potential to ribosome synthesis and critical cell
size. Genes & Development, 18(20):2491-2505, October 2004.
}
\keyword{datasets}
