\name{Sc.rpg.ensg}
\docType{data}
\alias{Sc.rpg.ensg}
\title{Ribosomal protein genes.}
\description{
ORF identifiers (Ensembl Gene ID) encoding for ribosomal protein genes.
}
\usage{Sc.rpg.ensg}
\format{
Vector of ORF identifiers (Ensembl Gene ID).
}
\source{
A. Nakao, M. Yoshihama, and N. Kenmochi. RPG: the Ribosomal Protein Gene database.
Nucleic acids research, 32(Database issue), January 2004.
}
\keyword{datasets}
