\name{Sc.stress.ensg}
\docType{data}
\alias{Sc.stress.ensg}
\title{ISA stress module.}
\description{
ORF identifiers (Ensembl Gene ID) found to be associated with stress response by the iterative signature algorithm.
}
\usage{Sc.stress.ensg}
\format{
Vector of ORF identifiers (Ensembl Gene ID).
}
\source{
J. Ihmels, G. Friedlander, S. Bergmann, O. Sarig, Y. Ziv, and N. Barkai. Revealing modular
organization in the yeast transcriptional network. Nature genetics, 31(4):370-377, August 2002.
}
\keyword{datasets}
