\name{Sc.tf.ensg}
\docType{data}
\alias{Sc.tf.ensg}
\title{Transcription factors.}
\description{
ORF identifiers (Ensembl Gene ID) encoding for transcription factors.
}
\usage{Sc.tf.ensg}
\format{
Vector of ORF identifiers (Ensembl Gene ID).
}
\source{
K. D. MacIsaac, T. Wang, D. B. Gordon, D. K. Gifford, G. D. Stormo, and E. Fraenkel. An
improved map of conserved regulatory sites for saccharomyces cerevisiae. BMC Bioinformatics,
7:113, 2006.
}
\keyword{datasets}
