\name{Sp.affy.reliable}
\docType{data}
\alias{Sp.affy.reliable}
\title{Gene identifiers valid for cDTA normalization from the Saccharomyces Cerevisiae Sun et al. cDTA experiment.}
\description{
Ensembl gene IDs, that passed certain criteria among the Saccharomyces Cerevisiae Sun et al. cDTA experiment to be considered valid for cDTA normalization. For details, see Sun et al (Materials and Methods).
}
\usage{Sp.affy.reliable}
\format{
Vector of Schizosaccharomyces Pombe affymetrix IDs that can be passed to \code{DTA.normalize} for cDTA normalization of the Saccharomyces Cerevisiae identifiers.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Under review.
}
\keyword{datasets}
