\name{Sun2011}
\docType{data}
\alias{Sun2011}
\title{Saccharomyces Cerevisiae rpb1-N488D (Slow Polymerase) and wild-type cDTA experiment from Sun et al.}
\description{
R object contains all relevant *.RData files needed for the \code{DTA.estimate} function. For example, see Schwalb et al.
}
\usage{Sun2011}
\format{
R object contains the following *.RData files:
\code{Raw.datamat}
\code{Sp.affy.reliable}
\code{Sc.affy2ensg}
\code{Wt.phenomat}
\code{Pol.phenomat}
\code{Sc.ensg.reliable}
\code{Sc.tnumber}
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Genome Research, in revision.
B. Schwalb, B. Zacher, S. Duemcke, D. Martin, P. Cramer, A. Tresch. Measurement of genome-wide RNA synthesis and decay rates with Dynamic Transcriptome Analysis (DTA/cDTA). Bioinformatics, in revision.
E. Birney, D. Andrews, M. Caccamo, Y. Chen, L. Clarke, G. Coates, T. Cox, F. Cunningham, V. Curwen, T. Cutts, T. Down, R. Durbin, X. M. Fernandez-Suarez, P. Flicek, S. Graef, M. Hammond, J. Herrero, K. Howe, V. Iyer, K. Jekosch, A. Kaehaeri, A. Kasprzyk, D. Keefe, F. Kokocinski, E. Kulesha, D. London, I. Longden, C. Melsopp, P. Meidl, B. Overduin, A. Parker, G. Proctor, A. Prlic, M. Rae, D. Rios, S. Redmond, M. Schuster, I. Sealy, S. Searle, J. Severin, G. Slater, D. Smedley, J. Smith, A. Stabenau, J. Stalker, S. Trevanion, A. Ureta-
Vidal, J. Vogel, S. White, C.Woodwark, and T. J. Hubbard. Ensembl 2006. Nucleic acids research, 34(Database issue), January 2006.
}
\keyword{datasets}
