\name{tDTAreliable}
\docType{data}
\alias{tDTAreliable}
\title{tDTA reliable ORFs}
\description{
S.cerevisiae genes, that passed certain criteria to be considered valid for parameter estimation. (see publication)
}
\usage{tDTAreliable}
\format{
Vector of ORF identifiers.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. submitted.
}
\keyword{datasets}
