\name{tDTAtnumber}
\docType{data}
\alias{tDTAtnumber}
\title{tDTA tnumber}
\description{
The amount of thymines in the cDNA of each transcript.
}
\usage{tDTAtnumber}
\format{
Vector gives the number of uridine residues for every ORF identifiers.
}
\source{
J. M. Cherry, C. Adler, C. Ball, S. A. Chervitz, S. S. Dwight, E. T. Hester, Y. Jia, G. Juvik,
T. Roe, M. Schroeder, S. Weng, and D. Botstein. Sgd: Saccharomyces genome database.
Nucleic Acids Res, 26(1):73-79, 1998.
}
\keyword{datasets}
