\name{Wt.phenomat}
\docType{data}
\alias{Wt.phenomat}
\title{Design of the Saccharomyces Cerevisiae wild-type cDTA experiment from Sun et al.}
\description{
The phenotype matrix \code{Wt.phenomat} contains information about the experimental design. It is comprised of the filename, the type of RNA fraction measured (T, U or L), the labeling time and the replicate number.
}
\usage{Wt.phenomat}
\format{
The phenomat is a matrix comprised of the file name, the type of RNA fraction mesasured (T, U or L, \code{fraction} column), the labeling time (\code{time},\code{timeframe} column) and the replicate number (\code{nr} column). Rows in this matrix represent the individual experiments.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Under review.
}
\keyword{datasets}
