\name{WTdatamat}
\docType{data}
\alias{WTdatamat}
\title{cDTA WTdatamat}
\description{
Gene expression profiles of wild-type cDTA.
}
\usage{WTdatamat}
\format{
This matrix contains the RNA intensity values for each gene across each RNA fraction and their replicate measurements. The column names of the matrix give the cel-file name and the row names the ORF identifier.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. submitted.
}
\keyword{datasets}
