\name{WTphenomat}
\docType{data}
\alias{WTphenomat}
\title{cDTA WTphenomat}
\description{
Design of a wild-type cDTA experiment.
}
\usage{WTphenomat}
\format{
The phenomat is comprised of the file name, the type of RNA fraction mesasured (T, U or L), the labeling time and the replicate number. Rows in this matrix represent the individual experiments.
}
\source{
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. submitted.
}
\keyword{datasets}
