\name{DTA-package}
\alias{DTA-package}
\alias{DTA}
\docType{package}
\title{Dynamic Transcriptome Analysis}
\description{
The \code{DTA} package implements all methods of the quantitative kinetic modeling approach belonging to DTA (Dynamic Transcriptome Analysis) to estimate mRNA synthesis and decay rates from individual time point measurements.
}
\details{
\tabular{ll}{
Package: \tab DTA\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2012-03-14\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{Bjoern Schwalb <schwalb@lmb.uni-muenchen.de>}
\references{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Doelken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.
M. Sun, B. Schwalb, D. Schulz, N. Pirkl, L. Lariviere, K. Maier, A. Tresch, P. Cramer. Mutual feedback between mRNA synthesis and degradation buffers transcript levels in a eukaryote. Under review.
B. Schwalb, B. Zacher, S. Duemcke, D. Martin, P. Cramer, A. Tresch. Measurement of genome-wide RNA synthesis and decay rates with Dynamic Transcriptome Analysis (DTA/cDTA). Bioinformatics.
}
\keyword{package}
\examples{
## see vignette or supplemental material of the given references.
}
