\name{DTA.estimate}
\alias{DTA.estimate}
\title{Estimation of synthesis and decay rates}
\description{DTA.estimate uses an experiment, given by a phenotype matrix, data matrix and the number of uridines for each gene to estimate synthesis and decay rate of the genes.}
\usage{
DTA.estimate(phenomat = NULL, datamat = NULL, tnumber = NULL, ccl = NULL, mRNAs = NULL, reliable = NULL, mediancenter = TRUE, usefractions = "LandT", LtoTratio = NULL, ratiomethod = "tls", largest = 5, weighted = TRUE, relevant = NULL, check = TRUE, regression = TRUE, labeling = TRUE, correctedlabeling = FALSE, rankpairs = TRUE, assessment = TRUE, correlation = TRUE, error = FALSE, bicor = TRUE, condition = "", upper = 700, lower = 500, plots = FALSE, notinR = FALSE, folder = NULL, addformat = NULL, totaloverwt = 1, simulation = FALSE, sim.object = NULL)
}
\arguments{
  \item{phenomat}{A phenotype matrix, containing the design of the experiment as produced by \code{DTA.phenomat}. Columns are name, fraction (U=unlabebeld, L=labeled, T=total), time and nr (=replicate number). Rows represent individual experiments.}
  \item{datamat}{A matrix, containing the measurements from U, L and T, according to the design given in phenomat. Matrix should only contain the rows of phenomat as columns.}
  \item{tnumber}{Integer vector, containing the numbers of uridines. Elements should have the rownames of datamat.}
  \item{ccl}{The cell cycle length of the cells.}
  \item{mRNAs}{Estimated number of mRNAs in a cell (optional).}
  \item{reliable}{Vector of 'reliable' genes, which are used for parameter estimation.}
  \item{mediancenter}{Should the quotient Labeled/Total resp. Unlabeled/Total be rescaled to a common median over it's replicates before building the genewise median.}
  \item{usefractions}{From which fractions should the decay rate be calculated: "LandT", "UandT" or "both".}
  \item{LtoTratio}{Coefficient to rescale Labeled/Total. Is estimated from the data, if not specified. See ratiomethod.}
  \item{ratiomethod}{Choose the regression method to be used, possible methods are: "tls", "bias" and "lm". For details, see the vignette.}
  \item{largest}{Percentage of largest residues from the first regression not to be used in the second regression step. For details, see the vignette.}
  \item{weighted}{Should the regression be weighted with 1/(Total^2 + median(Total))?}
  \item{relevant}{Choose the arrays to be used for halflives calculation, vector due to nr (=replicate number) in phenomat.}
  \item{check}{If check=TRUE, control messages and plots will be generated.}
  \item{regression}{Should the regression results be plotted?}
  \item{labeling}{Should the labeling bias be plotted?}
  \item{correctedlabeling}{Should the corrected labeling bias be plotted?}
  \item{rankpairs}{Should the ranks of decayrates be compared in pairs be plotted (based on Labeled/Total)?}
  \item{assessment}{Should 1-L/T, U/T or (1-L/T+U/T)/2 be assessed due to limitations of the decay rate formula?}
  \item{correlation}{Should the correlation of half-lives, decay rates and synthesis rates vs Total, Labeled and the number of uridines be plotted as color-coded squares?}
  \item{error}{Should the standard deviations and coefficients of variation be calculated?}
  \item{bicor}{Should the labeling bias be corrected?}
  \item{condition}{String, to be added to the plotnames.}
  \item{upper}{Upper bound for labeling bias estimation. For details, see the vignette.}
  \item{lower}{Lower bound for labeling bias estimation. For details, see the vignette.}
  \item{plots}{If plots=TRUE, control plots will be saved.}
  \item{notinR}{Should plots be not plotted in R.}
  \item{folder}{Path to the folder, where to save the plots.}
  \item{addformat}{Additional fileformat for plots to be saved. See \code{plotit} function (\code{LSD} package).}
  \item{totaloverwt}{Will be available in the very near future for comparative DTA data.}
  \item{simulation}{True, if data was generated by \code{DTA.generate}.}
  \item{sim.object}{Simulation object created by \code{DTA.generate}.}
}
\value{
\code{DTA.estimate} returns a list, where each entry contains the estimation results for all replicates of one labeling time. Each result contains the following entries
  \item{triples}{Mapping of each fraction and experiment to its corresponding column in the data matrix.}
  \item{plabel}{The labeling efficiency. For details, see the vignette.}
  \item{LtoTratio}{Estimated ratio of labeled to total fraction.}
  \item{UtoTratio}{Estimated ratio of unlabeled to total fraction.}
  \item{LtoUratio}{Estimated ratio of labeled to unlabeled fraction.}
  \item{correcteddatamat}{Labeling bias corrected data matrix.}
  \item{drmat}{Decay rates for each replicate. The last column gives the median decay rates.}
  \item{hlmat}{Half-lives for each replicate. The last column gives the median half-lifes.}
  \item{dr}{Median decay rates. The last column of drmat.}
  \item{hl}{Median half-lives. The last column of hlmat.}
  \item{dr.sd}{Standard deviations of decay rates.}
  \item{dr.cv}{Coefficients of variation of decay rates.}
  \item{hl.sd}{Standard deviations of half-lives.}
  \item{hl.cv}{Coefficients of variation of half-lives.}
  \item{TEmat}{Total expression for each replicate. The last column gives the median total expression values.}
  \item{TE}{Median total expression values. The last column of TEmat.}
  \item{TE.sd}{Standard deviations of total expression values.}
  \item{TE.cv}{Coefficients of variation of total expression values.}
  \item{LEmat}{Labeled expression for each replicate. The last column gives the median labeled expression values.}
  \item{LE}{Median labeled expression values. The last column of LEmat.}
  \item{LE.sd}{Standard deviations of labeled expression values.}
  \item{LE.cv}{Coefficients of variation of labeled expression values.}
  \item{UEmat}{Unlabeled expression for each replicate. The last column gives the median unlabeled expression values. (Only if unlabeled values exist in the experiment)}
  \item{UE}{Median unlabeled expression values. The last column of UEmat. (Only if unlabeled values exist in the experiment)}
  \item{UE.sd}{Standard deviations of unlabeled expression values.}
  \item{UE.cv}{Coefficients of variation of unlabeled expression values.}
  \item{srmat}{Synthesis rates for each replicate. The last column gives the median synthesis rates.}
  \item{sr}{Median synthesis rates. The last column of srmat.}
  \item{sr.sd}{Standard deviations of synthesis rates.}
  \item{sr.cv}{Coefficients of variation of synthesis rates.}
  \item{LTmat}{Labeled to total ratio for each replicate. The last column gives the median labeled to total ratios.}
  \item{LT}{Median labeled to total ratios. The last column of LTmat.}
  \item{UTmat}{Unlabeled to total ratio for each replicate. The last column gives the median unlabeled to total ratios.}
  \item{UT}{Median unlabeled to total ratios. The last column of UTmat.}
  \item{Rsrmat}{Rescaled synthesis rates for each replicate, if parameter \code{mRNAs} is specified. The last column gives the median synthesis rates.}
  \item{Rsr}{Rescaled median synthesis rates. The last column of Rsrmat.}
  \item{globaldrmat}{Decay rate for each replicate. Reciprocally weighted by the total expression. Last element contains (weighted) median decay rate.}
  \item{globaldr}{(Weighted) median decay rate.}        

}
\references{
C. Miller, B. Schwalb, K. Maier, D. Schulz, S. Duemcke, B. Zacher, A. Mayer, J. Sydow, L. Marcinowski, L. Dolken, D. E. Martin, A. Tresch, and P. Cramer. Dynamic transcriptome analysis measures rates of mRNA synthesis and decay in yeast. Mol Syst Biol, 7:458, 2011.

}
\author{Bjoern Schwalb \email{schwalb@lmb.uni-muenchen.de}}
\seealso{
\code{\link[LSD:heatscatter]{heatscatter}}, \code{\link[LSD:plotit]{plotit}}, \code{\link{tls}}
}
\examples{
dataPath = system.file("data", package="DTA")
load(file.path(dataPath, "datamat.RData"))
load(file.path(dataPath, "phenomat.RData"))
load(file.path(dataPath, "tnumber.RData"))
load(file.path(dataPath, "reliable.RData"))

res = DTA.estimate(phenomat,datamat,tnumber,ccl = 150,mRNAs = 60000,reliable = reliable,assessment = FALSE)

dev.off()
}
\keyword{methods}

